// frontend/Types/dataTypes.ts
export interface HistoryMatchType {
  opponent: string;
  yourScore: number;
  opponentScore: number;
  date: string;
  login: string;
  avatar: string;
  fullname: string;
}